requirements-dev.txt
pytest==7.3.1
apache-airflow-client==2.6.0
testcontainers==3.7.1
SQLAlchemy
s3fs
pyodbc==4.0.39
apache-airflow==2.6.0